AUTOPILOT: Disable AP timestamp projection
EKF estimator doesn't like inputs that are timestamped in the future wrt the computer clock. This occasionally happens due to the way AP timestamp projection works in concert with the new time sync mechanism. Solution is to trust (for now) that the computer clock won't drift too far from the autopilot time, and not project.

Also, removing a problematic line from the launch file.