NETWORK: Fixed problem with UDP port binding
It turns out we cannot bind to an IP *and* receive broadcast
traffic. Reworked options and cases for Socket class such
that whenever IPs are manually specified it implies we are
binding to a specific IP and not just a port. This will
unbreak normal operation and keep single-UAV SITL working,
though will likely need to be modified down the road.