TASKS: Created tasks package and run_once script
run_once.py slates tasks to be run at different points in time
after the launch has completed. At the moment, it only supports
running events once the autopilot bridge is up and fully running.

It currently sets the autopilot's ID based on the payload ID,
and starts rosbag (which is delayed so that the payload clock
can be set first by the autopilot).