DEPLOY: Compensated for lack of odroid-config
Newer ODroid media no longer comes with the odroid-config script
that runs upon first boot. To compensate, partition and filesystem
expansion has been rolled into a new script, deploy.sh, which is
run on the machine that sets up the eMMC or SD card. It also copies
odroid-installer.sh over to the card.

odroid-installer.sh now also disables lightdm (the X display manager).

Added the old odroid-config to the repo for posterity; this is not
intended to be used directly but referenced as needed.