FOLLOW_ME: Integration with swarm tracker
Implemented the follow me capability using the swarm tracker (vice
acs_pose and recv_pose topics).  It is implemented as a subclass of the
nodeable class so that it can be run as a node (follower_node.py) or
incorporated as an object in another node.

By default, the controller is "inactive" when it starts up so it will
not actually issue commands.  It can be activated and deactivated by
publishing a message (Boolean) to the controllers/follower_run topic
or calling the setActive method (this is the same paradigm as is used
by the waypoint sequencer).  Follow parameters (leader aircraft, follow
distance, offset angle from direct trail, altitude separation, etc.)
can be set by publishing a FormationOrder message to the
controllers/set_form_params topic or calling the object's reset method.