UTILS: Added unwieldy --lo-reverse option
To accommodate the alternative multi-payload solution from recent
commits, added an option to all utilities to "flip" the loopback
IPs. This will result in an error unless a) a different port is
specified (see --port option) _and_ repeater.py is running and is
configured for all these ports.

Needless to say, this option ought not stick around too long
(though it probably will).