SWARM_TRACKER: Implemented subswarm tracking
Added subswarm tracking to the swarm tracker.  Swarm tracker subscribes
to the network/recv_set_subswarm topic for the subswarm to which the
the local aircraft belongs.  Other aircraft subswarms are maintained
through the subswarm_id field of the SwarmVehicleState messages
received through the network/recv_pose topic.

NOTE:  ap_network_bridge (aerial.py) is currently sending 0s for the
subswarm, so this is not fully incorporated.  Within aerial.py, the
subswarm is set with a ROS parameter at launch, but it does not seem
to be updated when the aircraft is assigned to a new subswarm via
the ROS topic (as written, a network message might make the switch,
but the capability to switch without an order from the network should
be available as well).