NETWORK: Allow ROS-land to update subswarm ID
Changed recv_set_subswarm topic to update_subswarm, and both publish
and subscribe to it. The subscriber updates the socket value and
the ROS param; on a network-side change, the socket value and ROS
param are updated *and* a message is published. The subscriber does
nothing if the commanded value is the same as the current value,
preventing a loop from occurring.