PATH_PLANNING:  FollowController safety fix
Added a check to the FollowController object and follower_node.py to
make sure that an aircraft is not ordered to follow itself (which
would cause the aircraft to fly away over the horizon, failsafe
notwithstanding).  Within follower_node.py, if the aircraft ID and
follow ID are the same, the node will print an error message and
exit.  Within the object, an error message will be printed if the
object is reset to follow itself and the reset will not work (and
the controller will not activate), and error message will be
printed if an attempt is made to activate the controller when the
aircraft ID and follow ID are the same (to make it harder to screw
it up through bad programming).