PATH:  Controller launch file (disabled and uninitialized launch)
Added controllers.launch and controllers_ns.launch to launch all of
the controller nodes (currently just follwer_node.py and
wp_sequencer_node.py) in an inactive and uninitialized state.
controllers.launch launches all of the controllers in the base
namespace, while controllers_ns.launch allows for a namespace
argument (arg ns) to be compatible with the multi-sitl launch.
Both controllers have been modified to track their "is_active"
and "is_ready" states.  As written it is not possible (hopefully)
to change a node's is_active state (meaning that it is issuing
orders) if the is_ready state is false.

Both nodes are listening to ROS topics to receive new orders and
to be activated or can be set and activated directly using methods
(none of this is a change from the previous version).