MSGS:  FormationOrderStamped added
Added a FormationOrderStamped message to add a header field to the
FormationOrder message (the FormationOrder is embedded in the new
message in an order field).  Also, deleted the overshoot field from
the message (it's probably more appropriate for the controller to
compute this value) and changed the offset_angle field to angle.
The value of the angle field is now relative bearing from the lead
aircraft vice offset angle from directly astern the lead aircraft.