MISSION: Initial commit of ctlr_select.py
ctlr_select is a "Controller-Selector" node, that checks and
approves network commands to select from among the path planning
controllers. It attempts to perform enough sanity checks to
ensure that activating the requested controller will not cause
flight safety issues. It also attempts to ensure that at most
one controller is active at a time.

This is a work in progress; the code "runs" as-is but only in the
sense that it does not produce any errors upon startup. It has not
been tested *at all*. Note that its topics must be remapped in
order to be utilized.