PATH:  Altitude control in FollowController
Changed the altitude control in the FollowController class to treat the
altitude in the FormationOrder message as an absolute altitude (MSL) when
the altitude control mode is set to BASE_ALT_MODE.  The controller handles
the conversion to relative altitude prior to publishing the command to the
payload_waypoint topic so that they are in the form required by the autopilot.
Also, fixed a computation bug when using ALT_SEP_MODE (the old code would
use the lead aircraft alt field and add the separation to generate the
waypoint order, but needs to subtract __this__ aircraft's base altitude
so that the altitude is relative to the launch altitude).