PATH:  FollowController tweak to improve range performance
Modified the FollowController to compute the waypoint along the
lead aircraft's path instead of just through the target point.  The
waaypoint is now computed as the vector sum of 1 overshoot distance
from the follow aircraft to the target point and one overshoot
distance in the direction of lead aircraft travel.  The previous
version led to aircraft position stabilizing (more or less) about
100 meters behind the target point.  There's still a little bit
of lag, but it is substantially less in this version.