PATH and LIB:  Restructured Controller object (lib) and subclasses (path)
Moved common functionality from the WaypointSequencer and FollowController
classes to the Controller class.  Changing the is_ready status is now
initiated by the subclass (e.g., when a new control order is received)
but managed by the parent class.  Activation and deactivation of the
controller is now handled by the parent class.  The executeTimedLoop
method was also moved from the subclass to the parent class.  The method
now issues the status message (as required) and calls a subclass runController
method (virtual method of the Controller class) to generate control orders
if and only if the object is both "ready" and "active".  These changes make it
less likely that a controller that is not ready can be forced to issue
commands.