PATH:  Bug fix to waypoint controller
Did not correctly implement class hierarchy requirements for this node.
The executeTimedLoop method was still present, so it was overriding
the Controller class method.  Renaming the method to runController
fixes this (this is the virtual method from the Controller class that
is called by the executeTimedLoop function in the Controller class)