PATH:  Tracking error in follow_controller.py
Fixed a bug (that was introduced by my last bug fix) in follow_controller.py.
The forward biased behavior (i.e., projecting past the follow point along
the leader's path) works well normally, but if the follower is in front
of the leader (usually only when the controller is initially activated)
there's some weird behavior that can cause the aircraft to fly away
(not completely, but further than we want).