MISSION_PLANNING:  Waypoint services to ID loiter waypoints
Incorporated mavbridge waypoint services to identify infinite loiter
waypoints in the autopilot mission file.  The autopilot bridge service
is used to load all waypoints at startup and any time a controller
is to be activated when none currently are (i.e., when the autopilot
is solely responsible for control--this is to ensure that the most
current list is available in case they have been changed in the
autopilot since the last time the payload had control).  Each waypoint
in the file is checked, and a list of available infinite loiter
waypoints are identified (command == 17).  This isn't being used
for anything right now, but it gives us the availability to implement
a more robust safety mode.  The launch files were also modified to
load the wp module of the autopilot bridge.