PATH:  Altitude safety check to existing controllers
Added a safety check for minimum ordered altitude to both controllers.
For the WaypointSequencer, if an invalid relative altitude (< 50m
right now) is called for, the controller assumes a bad waypoint list,
so it sets both the is_ready and is_active values to false.  For the
follower controller, an invalid relative altitude will force a
deactivation of the controller (is_active = False) but will not
set the is_ready value to false (the condition might just mean that
the lead aircraft got low, so the controller itself might be fine).
The minimum altitude value is based on a new constant in controller.py.