MISSION_PLANNING: Safety waypoint implementation
Implemented the safety waypoint functionality for controller activation
and deactivation as discussed.  When payload control is initiated, the
current aircraft position (lat, lon, rel_alt) is saved.  When payload
control is disabled, this saved location is published to the
payload_waypoint topic.

Also removed the loiter_waypoint ROS parameter.  The infinite loiter
waypoint to use during payload control must now be the last waypoint in
the mission file.  If the last waypoint in the file is not an infininte
loiter waypoint, payload control will be inhibited.  The waypoints are
loaded each time waypoint control is initiated, so if the waypoint set
is changed by the ground station, the newest last waypoint will be used.

Also deleted the loiter waypoint ros parameter from the controller
launch files.