LIBS: Extended Message type with second seqnum
Determined that, like with TCP, it is necessary for all messages to
contain TWO sequence numbers - one representing the sender's current
position in the stream (here, highest-numbered reliably-sent message)
and another representing the sender's belief about the remote-end's
position in the stream. These are now named .msg_sseq and .msg_dseq,
respectively.

The reliability flag has been moved to one of the spare bits in the
'subswarm' byte, and the field has been renamed to .msg_fl_rel