LIB:  Controller activation by service
Implemented a service at the Controller level for activation and
deactivation of individual controllers.  The service uses the ap_srvs/
SetBoolean service to activate or deactivate the controller with the
response containing the new is_active value of the controller.  Service
requests are sent to nodename/nodename_run where "nodename" is the ROS
name of the controller's node.  The Controller object still prevents
activation of a controller that is not ready.  This implementation
does not preclude the use of topics to activate and deactivate
controllers, but that functionality will need to be implemented by
each controller.