LIB: SO_REUSEADDR and ignoring self as source
As it turns out, Linux is less restrictive about reusing UDP
sockets than it originally appeared:

http://stackoverflow.com/questions/14388706/socket-options-so-reuseaddr-and-so-reuseport-how-do-they-differ-do-they-mean-t

If all UDP sockets bound to the same IP and port have the option
SO_REUSEADDR set before the call to bind(), then things work. To
demonstrate this, it was also necessary to remove the check that
the source IP of incoming packets is our own (since it could be
in the multi-SITL use case). This is replaced with a post-parsing
check of the source ID. This *will* increase the computation for
all platforms slightly, since all messages broadcast from a host
will also get parsed and checked by that host. We can make this
behavior optional in the future if necessary.