MISSION_PLANNING:  Service for setting ctlr_selector mode
Implemented a service to change the active controller.  The service
uses the ap_srvs/SetInteger service and is available at
ctlr_selector/selector_mode.  The return value for the service will
be True if the requested mode was successfully activated and False
if it was not (all of the safety conditions must be met for a
controller to be activated, so a request for a certain control
mode does not equate to successful activation).