PATH_PLANNING:  Topic and service naming
Modified the naming of controller activation/deactivation services and
parameter setting topics to standardize the convention some.  Activation
and deactivation services are now named "'name'/'name'_run" and the
parameter setting topic is now named "controllers/'name'_set" where
'name' is the name of the node.  Also changed the name of the ctlr_selector
selector_mode topic and service to set_selector_mode to make it clearer.