LIB:  Auto standard topic and service naming
Implemented functionality in the Nodeable class to use standardize
the topic and service naming convention to use a "nodename/topicname"
convention.  Added createService, createPublisher, and createSubscriber
methods that will take care of this.  Should make programming a little
more straightforward since we won't have to worry at the programmer
level about naming (remapping can be done in the launch file).