UTILS: Added options to launch script
By default, launch script now starts up a payload in the default
network namespace (no containers) and on the default UDP port. It
still calculates the TCP port needed to connect to the SITL.

The default is fine, and easy, for a single payload instance. To
be more useful (and still easy to use) for multiple payloads,
try one of the following options:

-P : Uses calculated UDP port offsets (this is the old default),
so instance 1 will use 5555, instance 2 will use 5556, and so on.
This requires having a repeater (see repeater.py) to move packets
between instances.

-D DEVICE : Instead of using lo (loopback) in non-container mode,
use DEVICE (e.g., eth0). Using a device that supports broadcast
allows running multiple payload instances on the default UDP port
(5554). Note that in doing so, SITL+payload instances will be able
to talk to instances on other computers as well as actual planes.
Also note that, in office/lab settings, this will potentially
generate a lot of broadcast traffic on a network that might not
be happy to see it!