MSGS:  Change to the SwarmFormationOrder msg
Fixed one data type error in the SwarmFormationOrder message.
Deleted the subswarm_id parameter (the message only applies to this
message, so the command can only apply to the entire swarm or to this
vehicle's subswarm).  Added a Boolean "subswarm_only" field.  If this
value is True, the command applies only to the subswarm that this
vehicle belongs to; if it is False, the command applies to the whole swarm.