MISSION_PLANNING:  swarm_manager.py
Implemented a SwarmManager object (swarm_manager node) to handle
swarm commands at the aircraft level.  The intent is that this node
will ingest swarm commands, figure out what this particular aircraft
is supposed to do, set the appropriate controller parameters, and
direct the ctlr_selector node to activate the required controller
when the swarm behavior is activated.

A swarm formation command has been implemented so far that uses a
SwarmFormationOrderStamped message to define formation parameters for
the swarm (or subswarm).  Altitudes are used to determine which
aircraft follows which (low to high for positive separation, high to
low for negative separation).