PERCEPTION:  SwarmTracker and network fail
Implemented a timing check in SwarmTracker so that an aircraft that
fails to report over the network for a predetermined amount of time
(5.0 seconds at present) will be dropped from the swarm.  If another
aircraft is "following" the dropped aircraft, the follow controller
will automatically deactivate (and set is_ready to false).  This will
fix the observed follow behavior following leader or follower network
 node failure (the follower will proceed to the safety waypoint as
soon as the leader drops out of the swarm, which will occur upon
network node termination by either aircraft).