NETWORK: Initial commit of network.py
This is a direct replacement to aerial.py, better written, easier
to extend, and more robust to exceptions and errors. It has been
mildly tested, and should be tested much, much more before flying.

Note that some topic and service names have changed, and a subsequent
commit is necessary to bring the launch files into alignment.