MISSION_PLANNING:  ctlr_selector.py inheritance
Refactored ctlr_selector.py to inherit from Nodeable.  This will make
the logging and printing a little more robust and should help identify
why it occasionally shifts out of payload control immiately after it is
initiated.  The launch files required some remapping to support this
change, so that is also included in this commit.