MISSION_PLANNING:  ctlr_select.py threading
Added a couple of flag variables to account for threading issues that
cause a newly-activated controller to be immediately deactivated on
occasion.  The new variables will allow a single erroneous autopilot
status message or controller status message (for each controller) to
be received without causing the selector to deactivate everything.
This will prevent status messages that are received prior to the
controller activation but not processed until after the activation
from causing a reset.