COMMS: Minor bug fix, documentation
Fixed a (theoretical) bug where the ACK field was being reset for
packet resends, even if it was already set. This introduced the
possibility that a lost SYN message could not be resent (if the ACK
is set to None, it will not be sent). Commented out the offending
lines and changed logic for resetting a message's ACK. It is now
reset if (and only if) the ACK is set for that remote pair. This
should retain the ACK of a SYN message but not falsely set the ACK
of any other message.

Documented the protocol in abstract form for further analysis.