NETWORK:  Publisher latch
Changed the bridge publish method to take a "latched" parameter (default
False) so that the publisher can be set up with the latch parameter
set as required.  The individual calls to the method were modified
to account for the new parameter (my best guess as to which ones need
the value set to True).