MISSION_PLANNING:  SwarmMannager subswarm management
Reworked the swarm manager to use the swarm_uav_states message only
to manage swarm and subswarm membership (the subswarm_uav_states
message is no longer subscribed to).
NOTE:  The auto-follow computations have not been tested and will not
work!  We're not using that right now, and it will probably be
completely rewritten before we do anyway.