COMMS: SwarmReady message, FlightStatus bit
Structured like FlightReady message, but denotes a toggling of
the aircraft's readiness to be used as part of the swarm.

Also added a new flag field to the FlightStatus definitions, .swarming,
that should be wired to correspond to the "swarm readiness". It uses
up one more bits of the "mode/flags" field.