UTILS: Fixed timer issue, relaxed constraints
1. Fixed the CPU-pegging issue, which turned out to be due to a
timer problem. Replaced with a different timer variant and the
problem appears to have disappeared.

2. Relaxed constraints for various actions, particularly confirming
swarm ready and shutting down a payload. This is mostly to support
development and debugging, but also to handle corner cases where
the aircraft isn't in quite the "right" state when the action must
be performed.