NETWORK: Exception handling more careful
We were still seeing the network bridge die in the field. One of
the few remaining theories is that logging errors (in the outer
try-except block) were generating futher, uncaught exceptions.

Placed a try-except block _inside_ the outer except block that,
if logging causes an error, simply passes and re-enters the loop.