DEPLOY: Can enter date manually
For make and catkin_make to work correctly, the system clock needs
to be correct (set after the previous time make was run). We can
use NTP to do this, but if NTP is not accessible for some reason,
the user can enter a date string by hand. 'date' takes a variety
of formats, but one recommended and concise format is:

MMDDhhmmCCYY.ss

So March 3, 2015 at 10:08 AM would be: 030310082015.00