NETWORK: Replaced SwarmReady w CalPress, threads
Got rid of SwarmReady message and inclusion of swarming flag in
network status message. Added CalPress message to initiate
airspeed sensor calibration (CALibrate PRESSure).

Because we want feedback on pressure calibration, it is defined
(in mavbridge) as a service. It takes time, and we don't want to
hold up network message processing, so a new doInThread() method
was added. It takes a main callback (the work to do in a separate
thread), and an "error" callback (which is used if either the
thread setup or the main callback raises an exception).

This is used by CalPress to perform the pressure calibration,
await feedback (which influences a flag in the Status message),
and prevent restarting calibration while it is in progress.