MISSION_PLANNING:  Swarm Manager update
Fairly significant rework of swarm manager.  Swarm behaviors changed to a
registry system so that new ones can be added easily.  Moved most of the
functionality for computing aircraft requirements for a swarm behavior to
this node (what we've got was mostly in SwarmCommander previously).  The
intent is to provide a little cleaner framework for implementing new swarm
behaviors as they are developed.  Also changed the swarm state management
functionality to rely on things other than waypoint numbers (some of that
still remains for now, but has been marked with comments so that it can be
changed when we've got something better to work with).
And some other stuff, too.