PATH_PLANNING:  follow_controller intercept behavior
Modified the intercept behavior of the follow controller.  Capped the
"lookahead" to 10 seconds so that an intercepting vehicle that is a long
way from the lead aircraft won't project the leader's path too far into
the future (potentially driving off range to intercept a vehicle that was
going to change course long before it got to that point anyway).