COMMS:  New SwarmState message and mod to the FlightStatus message
Added a SwarmState message for the SwarmCommander to use to manually set the
swarm state of selected vehicles.  Under normal circumstances, UAVs determine
their swarm state based on sensed conditions, but this will allow manual reset
of the swarm state for testing purposes in cases where the UAV-determined
state needs to be reset to a different value for new tests.

Also, added the currently active swarm behavior to the FlightStatus message
(using a previously unused byte) so that it can be displayed in the
SwarmCommander dashboard.