MISSION_PLANNING: Manual swarm state setting & active swarm behavior reporting
Added functionality to allow for manual reset of the current swarm state.
This will not normally be required, but may be desirable during testing.
Added some functionality (ROS topics and such) to support adding the currently
active swarm behavior to the FlightStatus message so that it can be displayed
in the SwarmCommander dashboard.