NETWORK_BRIDGE and MASTER: Manual swarm state reset functionality
Added functionality to use the SwarmState message (from the SwarmCommander)
to manually reset the UAV's swarm state (functionality added to swarm_manager
and acs_messages in the last two commits).  Changes to master.launch are
required to correctly map the ROS topics.
Also added the currently active swarm behavior to the FlightStatus message
so that it can be displayed in the SwarmCommander dashboard.