MISSION_PLANNING:  New test for "landing" swarm state
Modified the test for the swarm manager to transition to the "landing" state
so that it uses the type of waypoint instead of a fixed waypoint number.
As new waypoints are executed, the waypoint data is pulled from the autopilot
using the wp_getrange service and the type of waypoint is cached.  Any time
the aircraft transitions to a "land here" waypoint (21), the swarm manager
will transition to the "landing" state.