MISSION_PLANNING: On board computation of swarm follow settings
Modified swarm manager to perform on-the-fly computation of swarm follow
parameters.  The most recently ordered waypoint's altitude is compared to
other subswarm aircraft altitudes, and the highest aircraft is chosen to
lead.  This eliminates the requirement for a hard-coded swarm follow
altitudes for all aircraft (they'll figure it out based on their altitudes
at the time the swarm follow is ordered).