MSGS:  Added SwarmControlState msg & modified SwarmVehicleState msg
Added a SwarmControlState message to provide a way for the swarm state
of other vehicles in the swarm (subswarm_id, swarm_state, swarm_behavior,
control_mode, and autopilot_mode) to be relayed to nodes from the network
node when autopilot status messages are received.

Modified the SwarmVehicleState message to include swarm control info:
swarm_state and swarm_behavior.