MISSION_PLANNING: Swarm manager subswarm assignment
Reworked subswarm assignment so that vehicles can be assigned to subswarms
from either the Swarm Ready or Egress state (was just Swarm Ready).  The
current swarm state is checked to make sure that no other vehicles with a
different swarm state are already assigned to that subswarm ID (this would
lead to potentially unsafe situations when a subswarm with vehicles in
different states are told to execute a particular swarm behavior).