MISSION_PLANNING: Swarm behavior activation changes
Modified the swarm behavior activation stuff some so that behaviors can
be available in either normal swarming or post-egress swarming (or both).
This will make it easier for behaviors to be implemented to work differently
in different portions of the flight (e.g., "standby" when in the "swarm ready"
or "swarm active" state will send the vehicle to the swarm standby waypoint,
but in the "egress" state will send it to the egress waypoing) or make
certain behaviors available only during specific portions of the flight
(e.g., "swarm follow" is available only when the vehicle is in the "swarm
ready" or "swarm active" state).