LIB:  Added enumerations file for project-wide constants
Moved some of the enumerated variables from various files to a single
ap_enumerations.py in the ap_lib package so that any node throughout the
project (including SwarmCommander or others) can access the same set of
enumerations and global declarations.  The intent is to reduce the likelihood
of mismatched values between packages as well as multiple names for
essentially the same value.