PATH_PLANNING:  swarm_landing_sequencer.py
First cut at a swarm landing sequencing capability.  The node determines
what UAV to follow (or whether this one is first) and goes to the landing
stack waypoint.  Altitude will stay 15m above the followed aircraft (no lower
than 100m) until the leader switches to "LANDING" (flare), at which time the
controller will shift this UAV to the "loiter_to_altitude" waypoint
immediately prior to the landing waypoint.

The landing waypoint sequence is hard coded for the moment to coincide with
the current ARSENL swarm_sortie_west mission (landing west to east) but
will be reworked once the swarm behavior messages are extended to support
parameters.  The sequence is based solely on aircraft altitudes for now and
does not use any sort of "negotiation", but the infrastructure is there to
allow for more robust sequencing algorithms.