MISSION_PLANNING: Added swarm_landing_sequencer to control structure
Added swarm_landing_sequencer controller to ctlr_selector and swarm_manager.
Also modified ctlr_selector.py to allow for controller types that do not
utilize the infinite loiter waypoint mechanism (the default is "True"
so no existing code needs to be modified).  This is required by the
swarm_landing_sequencer because the last thing it does is switch the waypoint
to the landing waypoint (which is not an infinite loiter waypoint and will
cause the previous version of ctlr_selector to automatically deactivate the
controller).  Going forward, we may utilize this capability more to implement
controllers that rely on something other than waypoint control.

Also fixed a bug in the ctlr_selector that caused the autopilot mode to
incorrectly be assessed as incompatible with the payload control in some
situations.