PATH_PLANNING: Improfed swarm_sequenced_landing behavior
Added parameters (landing waypoint sequence start) to the sequenced swarm
landing behavior.  Also changed it so that the orbit direction for the
infinite loiter waypoint used for most of the behavior is the same direction
as the loiter-to-altitude waypoint that sets up the landing (not doing this
can cause the UAV to miss the runway).