MISSION_PLANNING: Fully parameterized the fixed follow swarm behavior
Added functionality to the fixed follow swarm behavior (SWARM_FIXED_FORMATION)
to use all of the configuration parameters in the message.  The behavior will
now allow the user to specify range from the lead aircraft, angle-off-the-bow
from the lead aircraft, and whether following aircraft take the same position
relative to the leader (stacked) or relative to the next-higher aircraft in
the formation.