MSGS: Added an "is_paused" field to the controller status message
Added "is_paused" so that controllers can be paused without deactivating
them.  This might be particularly useful if we're doing something like
landing the aircraft and need to stop things for some reason without
sending the aircraft back to the swarm standby state (which might not
be safe if aircraft have already changed altitudes).