LIB: Pause functionality added to the Controller class
Added a "Pause" capability to the Controller class so that operation can be
interrupted without disabling the controller (meaning that it is still
active but not generating control commands).  This capability was requested
to allow the sequence landing behavior to be paused without all of the
UAVs being ordered back to the EGRESS state (which might involve unsafe
altitude changes), but can be used with other control modes as well.

Also added an ERROR value to the swarm state enumeration.  This value
probably indicates that the autopilot is not in AUTO mode (in which case
there is no "correct" swarm state).