COMMS: Added a PauseSwarmBehavior message
Added a PauseSwarmBehavior message to the message set.  This message differs
from the SuspendSwarmBehavior in that a behavior pause can be unpaused
and the previous behavior will resume from there (i.e., the behavior remains
active while paused) while a suspended behavior is fully deactivated.  This
capability was requested to allow the sequence landing behavior to be
interrupted without sending all of the UAVs back to the EGRESS state (which
would include potentially unsafe altitude changes).