PATH_PLANNING: Pause behavior and swarm landing safety checks
Modified the follow controller to correctly handle a pause (UAV will stop
following the designated lead aircraft and orbit in place).  Also added
safety feature to disable the controller if the lead aircraft autopilot
shifts out of AUTO (i.e., don't follow the lead to RTL).

Added safety functions to the swarm landing sequencer to make sure that the
UAV directly ahead of this one is in the correct mode.  If the followed
UAV drops out of the subswarm or isn't in the SWARM_SEQUENCE_LAND behavior,
the UAV recomputes its place in the landing order.