MISSION_PLANNING: Egress-to-Standby rtn to egress vice swarm ready point
Fixed a minor bug that caused the swarm state and swarm behavior to revert to
"swarm ready" and "standby" when the "terminate swarm behavior" button was
pressed (i.e., receives a SuspendSwarmBehavior network message).  Once the
aircraft has been egressed, the correct behavior is for it to remain in the
"egress" swarm state and return to the egress point.

also re-added Dylan's "return success" statement that was inadvertently
removed by Matt's rebase.