PATH_PLANNING: Autopilot notification & alt bug fix to swarm landing
Added functionality to the swarm landing sequencer to notify the autopilot
when a final descent for landing has been ordered or aborted so that the
autopilot can turn the fence on and off and conduct any other required
operations.  This functionality has been tested against the set_landing
branch in the autopilot_bridge package (subsequently merged into master).

Also changed how the initial altitude (transit to the landing stack) is
being ordered.  When the landing sequencer is initiated, it will use the
autopilot/wp_getrange service to retrieve the most recently ordered wpt.
The altitude from this waypoint (rel_alt) is used now.

Also added a minimum time (15sec) to loiter at the landing waypoint before
descending (will prevent early descent of the first aircraft).

Fixed a small bug in follow_controller ("and" to "or" in test to prevent
division by zero).

Also made changes to use project-wide enums vice local variables for
base altitude and altitude block size.