PATH_PLANNING: Send sequence land signal to the autopilot earlier.
The Sequence Lander needs to send the signal to the autopilot that it
has started a plane landing as soon at the START_TRANSIT phase, vice
just when the plane starts to descend.  Otherwise, the plane could
be sitting in a loiter waiting to land and then experience an
autoland failsafe.  The plane (already about to land) will restart
the landing sequence AND ALSO have no awareness of the peers that
are landing.

This fixes that issue.