MISSION_PLANNING: Controller architecture refactoring
COmbined functionality of the swarm_manager and ctlr_selector nodes into a
single swarm_control node.  The new node maintains a registry of all available
behaviors, fields activation request for individual behaviors, and controls
the activation and deactivation of all behaviors.

In this reorganization, behaviors are applied to the entire subswarm, so there
are no longer UAV-specific controllers that are activated and deactivated by a
swarm_manager (through a controller activation node) to realize individual
swarm behaviors.  Now each swarm behavior is its own discrete entity, so
individual UAV activaity within the swarm is determined by the behavior itself.