COMMS: SwarmBehaviorData mod to fix Python3 packing error
Added a hack based on type to fix an error when Python3 packed the byte array
in the SwarmBehaviorData message.  Python2 (payload) treats the bytes object
as a 'str' object, which has to use an ord() cmd to convert individual
elements to ints for packing in the message.  Python3 treats the same data
element as a 'bytes' object and throws an exception if the ord() method is
used on the individual elements.