PERCEPTION: Added an Eventually Reliable Crash Detector to swarm_tracker
Added an increasing window crash detector to swarm_tracker so that UAVs that
fail to report their position for an extended period will be suspected of
crashing.  The window size increases when a previously suspected crashed UAV
reports.  In testing with the lossy_comms branch (with 95% loss rate) the
window stabilizes at about 10 seconds.