MSGS: Added a crashed_list field to the SwarmStateStamped message
The field provides a list of UAV IDs for which no update information has
been recently received.  The records for those UAVs are still included in
the message, so it is up to the subscribing node how to treat the
potentially crashed UAVs (i.e., as part of the swarm or not).